requirements.txt
twisted==20.3.0 # fix for uvicorn/daphne error
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==7.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
uvicorn==0.11.8  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.0.10  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.42.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.11.1  # https://github.com/encode/django-rest-framework
django-cors-headers==3.5.0 # https://github.com/adamchainz/django-cors-headers
django-webpack-loader==0.6.0  # https://github.com/owais/django-webpack-loader
django-multiselectfield==0.1.12

# [OWN_DD]

django-filter==2.4.0
django-lockdown==3.0.0
django-simple-history==2.12.0
djangochannelsrestframework==0.2.0
django-rest-auth==0.9.5

channels==3.0.2
channels-demultiplexer==1.0.1
channels-redis==3.2.0
channelsmultiplexer==0.0.3

python-dateutil==2.8.1
django-background-tasks==1.2.5

drf-yasg==1.20.0

# for UML visualization

# pydotplus==2.0.2
# django-extensions==3.1.1

Werkzeug==1.0.1 # https://github.com/pallets/werkzeug
ipdb==0.13.3  # https://github.com/gotcha/ipdb
psycopg2==2.8.6  # https://github.com/psycopg/psycopg2
watchgod==0.6  # https://github.com/samuelcolvin/watchgod

# Testing
# ------------------------------------------------------------------------------
mypy==0.770  # https://github.com/python/mypy
django-stubs==1.5.0  # https://github.com/typeddjango/django-stubs
pytest==6.0.2  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.4  # https://github.com/Frozenball/pytest-sugar

# Documentation
# ------------------------------------------------------------------------------
sphinx==3.2.1  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2020.9.1 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8==3.8.3  # https://github.com/PyCQA/flake8
flake8-isort==4.0.0  # https://github.com/gforcada/flake8-isort
coverage==5.3  # https://github.com/nedbat/coveragepy
black==20.8b1  # https://github.com/ambv/black
pylint-django==2.3.0  # https://github.com/PyCQA/pylint-django
pre-commit==2.7.1  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.0.1  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==2.2  # https://github.com/jazzband/django-debug-toolbar
# django-extensions==3.0.8  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.8.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.9.0  # https://github.com/pytest-dev/pytest-django

sphinx-rtd-theme==0.5.1 # styling of the documentation

graphviz==0.16
pydotplus==2.0.2
django-extensions==3.1.1


gunicorn==20.0.4  # https://github.com/benoitc/gunicorn
psycopg2==2.8.6  # https://github.com/psycopg/psycopg2
sentry-sdk==0.17.5  # https://github.com/getsentry/sentry-python

# Django
# ------------------------------------------------------------------------------
django-anymail==8.0  # https://github.com/anymail/django-anymail
